Create README.md
As computer vision continues to advance quickly, there is an increasing need for humanmachine interaction across various domains. Conventional input devices, such as keyboards and 
mouse, have been fundamental in facilitating human-computer interaction. However, with the advancement of technology, hand gestures have emerged as a new and innovative means of 
interacting with computers. This project introduces a system for recognizing hand gestures that allows individuals with disabilities to control machines and complete tasks based on their hand 
movements. The project discusses the challenges that people with disabilities face in interacting with machines and the potential benefits of using hand gestures as a means of control. The 
system incorporates a machine learning algorithm that identifies hand gestures and translates them into commands that can be sent to a machine. The system is low-cost, user-friendly, and 
adaptable to various situations and environments, and can be employed to manage a wide range of machines, including computers, robots, and home appliances. The system has the potential to 
boost the autonomy and quality of life of individuals with disabilities by giving them a means of more effectively controlling their environment. It can also benefit individuals without disabilities 
by providing a more natural and intuitive way of interacting with machines. The paper outlines the design and implementation of the hand gesture recognition system, as well as experimental 
results that demonstrate its efficiency and potential applications.